sheets_sync/requirements.txt
fastapi
uvicorn
pandas
numpy
openai
anthropic
python-dotenv
pydantic
email-validator
requests
chromadb
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
python-multipart
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
